apps/mac-client/python/requirements.txt
pynput==1.7.7
faster-whisper==1.1.1
mlx-audio>=0.3.1
opencc-python-reimplemented>=0.1.7
requests==2.32.4
urllib3==1.26.20
